requirements.txt
sqlglot==28.1.0
PyYAML==6.0.3